requirements.txt
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
numpy==1.16.2
pandas==0.24.2
pyecharts-jupyter-installer==0.0.3
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
SQLAlchemy==1.3.1
virtualenv==16.4.3
visitor==0.1.3
Werkzeug==0.15.1
WTForms==2.2.1
